requirements.txt
disnake==2.6.1
python-dotenv==0.21.0
qrcode~=7.3.1
pyzbar~=0.1.9
Pillow~=9.2.0
